Refactor how optimized methods for Gaudi replace the equivalent Transformers' methods
