+++ optimum/habana/models/__init__.py
@@ -1,1 +1,3 @@
-from . import gpt2, t5, vit
+from .albert import gaudi_albert_forward
+from .modeling_all_models import gaudi_get_extended_attention_mask, gaudi_invert_attention_mask
+from .vit import gaudi_vit_self_attention_forward
