+++ optimum/habana/models/albert/__init__.py
@@ -1,0 +1,1 @@
+from .modeling_albert import gaudi_albert_forward
