+++ optimum/habana/models/modeling_all_models.py
@@ -1,0 +1,80 @@
+from typing import Tuple
+
+import torch
+
+from habana_frameworks.torch.hpex import hmp
+from transformers.modeling_utils import ModuleUtilsMixin
+
+
+def gaudi_invert_attention_mask(self, encoder_attention_mask: torch.Tensor) -> torch.Tensor:
+    """
+    Same as https://github.com/huggingface/transformers/blob/a9eee2ffecc874df7dd635b2c6abb246fdb318cc/src/transformers/modeling_utils.py#L640
+    except that HMP is disabled for computing:
+        encoder_extended_attention_mask = (1.0 - encoder_extended_attention_mask) * torch.finfo(self.dtype).min
+    """
+    if encoder_attention_mask.dim() == 3:
+        encoder_extended_attention_mask = encoder_attention_mask[:, None, :, :]
+    if encoder_attention_mask.dim() == 2:
+        encoder_extended_attention_mask = encoder_attention_mask[:, None, None, :]
+    # T5 has a mask that can compare sequence ids, we can simulate this here with this transposition
+    # Cf. https://github.com/tensorflow/mesh/blob/8d2465e9bc93129b913b5ccc6a59aa97abd96ec6/mesh_tensorflow
+    # /transformer/transformer_layers.py#L270
+    # encoder_extended_attention_mask = (encoder_extended_attention_mask ==
+    # encoder_extended_attention_mask.transpose(-1, -2))
+    # HMP is disabled because (1.0 - encoder_extended_attention_mask) may be converted into bf16
+    # while torch.finfo(self.dtype).min is the min value of fp32, which leads to NaNs
+    with hmp.disable_casts():
+        encoder_extended_attention_mask = encoder_extended_attention_mask.to(dtype=self.dtype)  # fp16 compatibility
+        encoder_extended_attention_mask = (1.0 - encoder_extended_attention_mask) * torch.finfo(self.dtype).min
+
+    return encoder_extended_attention_mask
+
+
+def gaudi_get_extended_attention_mask(
+    self, attention_mask: torch.Tensor, input_shape: Tuple[int], device: torch.device = None, dtype: torch.float = None
+) -> torch.Tensor:
+    """
+    Same as https://github.com/huggingface/transformers/blob/a9eee2ffecc874df7dd635b2c6abb246fdb318cc/src/transformers/modeling_utils.py#L692
+    except that HMP is disabled for computing:
+        extended_attention_mask = (1.0 - extended_attention_mask) * torch.finfo(dtype).min
+    """
+    if dtype is None:
+        dtype = self.dtype
+
+    if not (attention_mask.dim() == 2 and self.config.is_decoder):
+        # show warning only if it won't be shown in `create_extended_attention_mask_for_decoder`
+        if device is not None:
+            warnings.warn(
+                "The `device` argument is deprecated and will be removed in v5 of Transformers.", FutureWarning
+            )
+    # We can provide a self-attention mask of dimensions [batch_size, from_seq_length, to_seq_length]
+    # ourselves in which case we just need to make it broadcastable to all heads.
+    if attention_mask.dim() == 3:
+        extended_attention_mask = attention_mask[:, None, :, :]
+    elif attention_mask.dim() == 2:
+        # Provided a padding mask of dimensions [batch_size, seq_length]
+        # - if the model is a decoder, apply a causal mask in addition to the padding mask
+        # - if the model is an encoder, make the mask broadcastable to [batch_size, num_heads, seq_length, seq_length]
+        if self.config.is_decoder:
+            extended_attention_mask = ModuleUtilsMixin.create_extended_attention_mask_for_decoder(
+                input_shape, attention_mask, device
+            )
+        else:
+            extended_attention_mask = attention_mask[:, None, None, :]
+    else:
+        raise ValueError(
+            f"Wrong shape for input_ids (shape {input_shape}) or attention_mask (shape {attention_mask.shape})"
+        )
+
+    # Since attention_mask is 1.0 for positions we want to attend and 0.0 for
+    # masked positions, this operation will create a tensor which is 0.0 for
+    # positions we want to attend and -10000.0 for masked positions.
+    # Since we are adding it to the raw scores before the softmax, this is
+    # effectively the same as removing these entirely.
+    # HMP is disabled because (1.0 - encoder_extended_attention_mask) may be converted into bf16
+    # while torch.finfo(self.dtype).min is the min value of fp32, which leads to NaNs
+    with hmp.disable_casts():
+        extended_attention_mask = extended_attention_mask.to(dtype=dtype)  # fp16 compatibility
+        extended_attention_mask = (1.0 - extended_attention_mask) * torch.finfo(dtype).min
+
+    return extended_attention_mask
